Made NS work with MAZE
Changed the function obs formatting to just extractor of BS, cuz in the end I was not using the observations for anything else than the BS calculation. Also, now the NN agent works by using the sensor readings in the maze. Still have to change all the other baselines and evolver